% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{calculate_ttm_density}
\alias{calculate_ttm_density}
\title{calculate_ttm_density}
\usage{
calculate_ttm_density(typeTopicMatrix)
}
\arguments{
\item{typeTopicMatrix}{type topic matrix, obtained by 'get_type_topics(lda)'}
}
\description{
Calculate the density (sparsity) of the type topic matrix
}

