% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{create_lda_dataset}
\alias{create_lda_dataset}
\title{create_lda_dataset}
\usage{
create_lda_dataset(doclines, ldaconfig, stoplist_fn = "stoplist.txt")
}
\description{
Create an LDA dataset from existing string vector
}

