% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{extract_doc_lengths}
\alias{extract_doc_lengths}
\title{extract_doc_lengths}
\usage{
extract_doc_lengths(instances)
}
\arguments{
\item{instances}{the dataset to query}
}
\description{
Extract the the number of tokens in each document. Same order as the documents
}
