% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{extract_term_counts}
\alias{extract_term_counts}
\title{extract_term_counts}
\usage{
extract_term_counts(instances)
}
\arguments{
\item{instances}{the dataset to query}
}
\description{
Extract how many times each word occurs. Same order as the vocabulary
}
