% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{extract_vocabulary}
\alias{extract_vocabulary}
\title{extract_vocabulary}
\usage{
extract_vocabulary(alphabet)
}
\arguments{
\item{alphabet}{Java MALLET Alphabet object, obtained by 'get_alphabet(lda)'}
}
\description{
Extract the vocabulary as a string vector from an MALLET Alphabet
}

