% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{get_log_likelihood}
\alias{get_log_likelihood}
\title{get_log_likelihood}
\usage{
get_log_likelihood(lda)
}
\arguments{
\item{lda}{the trained lda model}
}
\description{
Extract the log likelihood for each iteration
}
