% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{get_theta_estimate}
\alias{get_theta_estimate}
\title{get_theta_estimate}
\usage{
get_theta_estimate(lda)
}
\arguments{
\item{lda}{LDA sampler object}
}
\description{
Get an estimate of the document topic distribution
(the theta matrix) from an LDA sampler
}
