% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{get_topwords}
\alias{get_topwords}
\title{get_topwords}
\usage{
get_topwords(lda, nr_words = 20)
}
\arguments{
\item{lda}{LDA sampler object}

\item{nr_words}{number of top words per topic to retrieve}
}
\description{
Get the top words per topic from an LDA sampler
}

