% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{load_lda_sampler}
\alias{load_lda_sampler}
\title{load_lda_sampler}
\usage{
load_lda_sampler(ldaconfig, ds, store_dir = "stored_samplers")
}
\arguments{
\item{ldaconfig}{LDA config object}

\item{ds}{LDA dataset}

\item{store_dir}{directory name containing stored sampler}
}
\description{
Load an LDA sampler from file.
}
