% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{print_top_words}
\alias{print_top_words}
\title{print_top_words}
\usage{
print_top_words(word_matrix)
}
\arguments{
\item{word_matrix}{top word matrix}
}
\description{
Print the 'top words' from a sampled word matrix
obtained using 'get_topwords(lda)'
}

