% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{sample_pclda}
\alias{sample_pclda}
\title{sample_pclda}
\usage{
sample_pclda(ldaconfig, ds, iterations = 2000,
  samplerType = "cc.mallet.topics.PolyaUrnSpaliasLDA")
}
\description{
Run the PCLDA (default Polya Urn) sampler
}

