% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{sample_pclda_continue}
\alias{sample_pclda_continue}
\title{sample_pclda_continue}
\usage{
sample_pclda_continue(lda, iterations = 2000)
}
\arguments{
\item{lda}{LDA sampler}

\item{iterations}{how many iterations to sample}
}
\description{
continue sampling using a trained sampler
}
