% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{calculate_ttm_density}
\alias{calculate_ttm_density}
\title{calculate_ttm_density}
\usage{
calculate_ttm_density(typeTopicMatrix)
}
\description{
Calculate the density (sparsity) of the type topic matrix
}

