% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{create_lda_dataset}
\alias{create_lda_dataset}
\title{create_lda_dataset}
\usage{
create_lda_dataset(train, test = NULL, stoplist_fn = "stoplist.txt")
}
\arguments{
\item{train}{string vector with document data}

\item{test}{string vector with test document data}

\item{stoplist_fn}{filiename of stoplist file}
}
\description{
Create an LDA dataset from existing string vector. Each entry
in the vector must be a string with the following format:
<unique id>\\t<doc class>\\t<document content>
The document class is not used in by the LDA sampler.
The document content CAN have \\t in it.
}
