% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{new_simple_lda_config}
\alias{new_simple_lda_config}
\title{new_simple_lda_config}
\usage{
new_simple_lda_config(dataset_fn, nr_topics = 20, alpha = 0.01,
  beta = (nr_topics/50), iterations = 2000, rareword_threshold = 10,
  scheme = "polyaurn", stoplist_fn = "stoplist.txt", topic_interval = 10,
  tmpdir = "/tmp")
}
\description{
Create a new LDA config file
}

