% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{new_simple_lda_config}
\alias{new_simple_lda_config}
\title{new_simple_lda_config}
\usage{
new_simple_lda_config(dataset_fn, nr_topics = 20, alpha = 0.01,
  beta = (nr_topics/50), iterations = 2000, rareword_threshold = 10,
  stoplist_fn = "stoplist.txt", topic_interval = 10, tmpdir = "/tmp")
}
\arguments{
\item{dataset_fn}{filename of dataset (in LDA format)}

\item{nr_topics}{number of topics to use}

\item{alpha}{symmetric alpha prior}

\item{beta}{symmetric beta prior}

\item{iterations}{number of iterations to sample}

\item{rareword_threshold}{min. number of occurences of a word to be kept}

\item{stoplist_fn}{filenname of stoplist file (one word per line) (default "stoplist.txt")}

\item{topic_interval}{how often to print topic info during sampling}

\item{tmpdir}{temporary directory for intermediate storage of logging data (default "tmp")}
}
\description{
Create a new LDA config file
}
