% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcldar.R
\name{sample_pclda}
\alias{sample_pclda}
\title{sample_pclda}
\usage{
sample_pclda(ldaconfig, ds, iterations = 2000,
  samplerType = "cc.mallet.topics.PolyaUrnSpaliasLDA", testset = NULL,
  save_sampler = TRUE)
}
\arguments{
\item{ldaconfig}{LDA config object}

\item{ds}{LDA dataset}

\item{iterations}{number of iterations to run}

\item{samplerType}{Java class of the sampler. Must implement the LDASampler interface}

\item{testset}{If give, the left-to-right held out log likelihood will be calculated on this dataset}
}
\description{
Run the PCLDA (default Polya Urn) sampler
}
